# spec/openapi.yaml
openapi: 3.1.0
info:
  title: tifl API
  version: 0.0.1
  description: >
    Canonical contract for the tifl API server. The OpenAPI spec is the only
    contract between clients (browser / Tauri / Capacitor) and the server; every
    client is generated or hand-written against it and reaches no server
    internals. Endpoints are added here as their handlers land.
    See context/backend-server.md and context/architecture-overview.md.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /languages:
    get:
      summary: List enabled language plugins
      operationId: listLanguages
      responses:
        "200":
          description: Registered languages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Language"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /llm/models:
    get:
      summary: List available LLM models
      description: >
        Returns safe model metadata from the configured LLM gateway. Provider
        credentials stay on the gateway process and are never exposed to the
        client.
      operationId: listLLMModels
      responses:
        "200":
          description: Available gateway models
          content:
            application/json:
              schema:
                type: object
                properties:
                  models:
                    type: array
                    items:
                      $ref: "#/components/schemas/LLMModel"
                required: [models]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          $ref: "#/components/responses/BadGateway"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /auth/register:
    post:
      security: []
      summary: Register and immediately authenticate a user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "201":
          description: Account created; rotated refresh credential set as an HttpOnly cookie
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/login:
    post:
      security: []
      summary: Authenticate with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: Authenticated; refresh credential set as an HttpOnly cookie
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/refresh:
    post:
      security: []
      summary: Rotate the refresh cookie and issue a new access token
      operationId: refresh
      responses:
        "200":
          description: Refreshed
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      security: []
      summary: Revoke the current refresh cookie
      operationId: logout
      responses:
        "204":
          description: Logged out
          headers:
            Set-Cookie:
              schema: { type: string }
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/logout-all:
    post:
      summary: Revoke every refresh session for the current user
      operationId: logoutAll
      responses:
        "204":
          description: All refresh sessions revoked
          headers:
            Set-Cookie:
              schema: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/me:
    get:
      summary: Return the authenticated user
      operationId: me
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /profile:
    get:
      summary: Return the current user's profile and client preferences
      description: >
        Returns the typed product profile used for default session generation
        and UI bootstrapping. In local/no-auth mode this is the synthetic
        `local` user's profile. Missing fields are defaulted server-side.
      operationId: getProfile
      responses:
        "200":
          description: Current profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      summary: Partially update the current user's profile
      description: >
        Updates typed profile fields and shallow-merges the `preferences` object
        into the existing arbitrary client-preferences map. A preference value of
        `null` deletes that key. Unknown top-level fields are rejected so product
        state stays explicit; client-owned arbitrary values belong under
        `preferences`.
      operationId: patchProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfilePatch"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /skills:
    get:
      summary: Return the current user's skill tree for a language
      description: >
        Returns every skill defined for the requested language, grouped by
        category and left-joined with the current user's XP/tier progress.
        Untouched skills are included with tier 0 and 0 XP so the client can
        render the complete read-only competency map. If `language` is omitted,
        the current profile's active_language is used.
      operationId: listSkills
      parameters:
        - name: language
          in: query
          required: false
          schema:
            type: string
          description: Optional enabled language code; defaults to profile.active_language.
      responses:
        "200":
          description: Skill tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillTree"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /sessions:
    get:
      summary: List sessions for the current user
      description: >
        Newest-first session overview page for home/resume screens. Includes
        generation status, story_id when available, selected target/new counts,
        and compact task progress counts without opening the SSE stream or
        fetching task bodies.
        By default archived sessions are excluded. Pass archived=true to list
        archived sessions for restore/delete workflows. Pass language to scope
        the page to one registered language; omit it to list all languages.
      operationId: listSessions
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: archived
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: language
          in: query
          required: false
          description: registered language code to scope the page to; omitted lists all languages
          schema:
            type: string
      responses:
        "200":
          description: Session overview page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /sessions/generate:
    post:
      summary: Start generating a session (story + tasks)
      description: >
        Creates a session and kicks off the staged generation pipeline
        asynchronously, returning immediately. If language is omitted, the
        current profile's active_language is used. If level is omitted, the
        server derives it from verified skill tiers when the language provides
        deterministic level rules; otherwise the current profile level is used.
        Poll progress via /sessions/{id}/events. Returns 503 if no LLM gateway
        is configured.


        Topic-guided sessions (session_type=topic_guided) require a non-empty
        topic (400 otherwise) and run a scope_check stage first: an out-of-scope
        topic fails generation with error_code GEN_SCOPE_REJECTED and a
        human-readable error_detail rather than silently producing a
        system-driven story. System-driven sessions choose a topic automatically.
      operationId: generateSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRequest"
      responses:
        "202":
          description: Generation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionRef"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
  /sessions/{id}:
    get:
      summary: Get session detail and persisted generation progress
      description: >
        Returns one tenant-scoped session with the metadata needed to navigate to
        reader/tasks, retry failed generation, or render a progress screen without
        subscribing to SSE. Another user's session returns 404.
      operationId: getSessionDetail
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "200":
          description: Session detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      summary: Hard-delete a session and its owned generated content
      description: >
        Removes the caller's session plus session-owned generated content such as
        generation stages, tasks, phrase sets, story tokens and glossary entries.
        User knowledge and reader exposure history are preserved; if reader
        events reference the generated story, the story row remains as their
        historical anchor.
      operationId: deleteSession
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "204":
          description: Session deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /sessions/{id}/debug:
    get:
      summary: Get session debug timeline and LLM audit calls
      description: >
        Returns the same tenant-scoped session detail and generation-stage
        timeline as GET /sessions/{id}, plus llm_calls rows for that
        session/current-user pair ordered by call time. Another user's session
        returns 404. This bounded slice is authenticated-user scoped; broader
        dev/admin-only gating is a follow-up.
      operationId: getSessionDebug
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "200":
          description: Session debug detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDebug"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /sessions/{id}/archive:
    post:
      summary: Archive a session
      description: >
        Soft-archives the caller's session so it disappears from the default
        session list while remaining restorable from archived=true.
      operationId: archiveSession
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "204":
          description: Session archived
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      summary: Restore an archived session
      description: >
        Clears archived_at so the caller's session returns to the default session
        list.
      operationId: unarchiveSession
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "204":
          description: Session restored
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /sessions/{id}/content:
    get:
      summary: Get a session's content (story reference or phrase set)
      description: >
        Loads whatever a session produced, discriminated by content_type. For a
        story it returns a reference the client loads through GET /stories/{id}
        (story content_type); for an expression-guided phrase session it returns
        the phrase items inline (phrase_set content_type). Phrase sets are not
        tokenized, so the reader's story endpoints do not apply to them. A story
        session whose story has not been generated yet returns 404; another user's
        session returns 404.
      operationId: getSessionContent
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "200":
          description: Session content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /sessions/{id}/events:
    get:
      summary: Stream generation progress (SSE)
      description: >
        Server-Sent Events stream of stage transitions. On connect the current
        persisted stage state is replayed, then live events follow until a
        terminal `{"stage":"done"}` event. The terminal event is enriched from
        persisted session detail so late subscribers receive the same navigation
        fields as live subscribers: final status, session_id, story_id when
        ready, compact task progress, stage summary, and failed stage/error code
        when generation failed. token_rate is an approximate upstream
        tokens/second during story generation (a ticker animation); story text
        is never streamed.
      operationId: sessionEvents
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/GenerationEvent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /sessions/{id}/reading:
    post:
      summary: Mark a session as reading (ready→reading)
      description: >
        Transitions the session status from ready to reading and records
        reading_started_at. Idempotent: if the session is already reading or
        complete the call succeeds with 204.
      operationId: startReading
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "204":
          description: Session is now in reading state
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /sessions/{id}/complete:
    post:
      summary: Mark a session as complete (reading→complete)
      description: >
        Transitions the session status from reading to complete and records
        completed_at. Idempotent: if the session is already complete the call
        succeeds with 204.
      operationId: completeSession
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "204":
          description: Session is now in complete state
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /sessions/{id}/retry:
    post:
      summary: Retry a failed generation from its failed stage
      description: >
        Resumes generation using session_generation_stages: a full re-run if the
        story never completed, or only the failed task-type stages otherwise.
      operationId: retrySession
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "202":
          description: Retry started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionRef"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /stories/import:
    post:
      summary: Import text as a reader story
      description: >
        Persists caller-provided target-language text as a tokenized story the
        normal reader can load via GET /stories/{id}. Clients may send pasted
        text as JSON or upload one UTF-8 text/plain .txt file with
        multipart/form-data. PDF/EPUB extraction is deferred. If language or
        level is omitted, the current profile defaults are used. Imported stories
        are standalone reader content and do not create generation sessions or
        tasks.
      operationId: importStory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportStoryRequest"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ImportStoryUploadRequest"
      responses:
        "201":
          description: Imported story
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportStoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Language:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        key_strategy:
          type: string
          enum: [surface, lemma, root, stem]
        enabled: { type: boolean }
      required: [code, name, key_strategy, enabled]
...
    LLMModel:
      type: object
      properties:
        id:
          type: string
          description: Provider model id.
        name:
          type: string
          description: Display name when provided by the upstream.
        description:
          type: string
          description: Upstream model description.
        context_length:
          type: integer
          description: Maximum context length when provided by the upstream.
      required: [id]
...
  responses:
...
    Unauthorized:
      description: Authentication failed
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    BadGateway:
      description: Upstream failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"